ADD: DefaultMenu for navigating the embeds
