+++ README.md
@@ -21,39 +21,42 @@ from pretty_help import PrettyHelp
 bot = commands.Bot(command_prefix="!", help_command=PrettyHelp())
 ```
 
+
+
 ### Added Optional Args
 
-- `active_time` - Set the time (in seconds) that the message will be active default is 30s
-- `menu` - Optionally set a custom function for pagination to use instead of the built-in menu
 - `color` - Set the default embed color
 - `ending_note` - Set the footer of the embed. Ending notes are fed a `commands.Context` (`ctx`) and a `PrettyHelp` (`help`) instance for more advanced customization.
 - `index_title` - Set the index page name default is *"Categories"*
-- `navigation` - Set the emojis that will control the help menu. Uses a `pretty_help.Navigation()` instance.
+- `menu` - set a custom menu for navigating pages. Uses a `pretty_help.PrettyMenu()` instance. Default is `pretty_help.DefaultMenu()`
 - `no_category` - Set the name of the page with commands not part of a category. Default is "*No Category*"
 - `sort_commands` - Sort commands and categories alphabetically
 - `show_index` - Show the index page or not
+
+### pretty_help.DefaultHelp args
+
+- `active_time` - Set the time (in seconds) that the message will be active. Default is 30s
+- `page_left` - The emoji to use to page left
+- `page_right` - The emoji to use to page right
+- `remove` - The emoji to use to remove the help message
 
 
 By default, the help will just pick a random color on every invoke. You can change this using the `color` argument:
 
 ```python
-import discord
+
 from discord.ext import commands
-from pretty_help import PrettyHelp, Navigation
+from pretty_help import DefaultMenu, PrettyHelp
 
+# ":discord:743511195197374563" is a custom discord emoji format. Adjust to match your own custom emoji.
+menu = DefaultMenu(page_left="\U0001F44D", page_right="👎", remove=":discord:743511195197374563", active_time=5)
 
+# Custom ending note
+ending_note = "The ending note from {ctx.bot.user.name}\nFor command {help.clean_prefix}{help.invoked_with}"
 
 bot = commands.Bot(command_prefix="!")
 
-# custom ending note using the command context and help command formatters
-ending_note = "The ending not from {ctx.bot.user.name}\nFor command {help.clean_prefix}{help.invoked_with}"
-
-# ":discord:743511195197374563" is a custom discord emoji format. Adjust to match your own custom emoji.
-nav = Navigation(":discord:743511195197374563", "👎", "\U0001F44D")
-color = discord.Color.dark_gold()
-
-bot.help_command = PrettyHelp(navigation=nav, color=color, active_time=5, ending_note=ending_note)
-
+bot.help_command = PrettyHelp(menu=menu, ending_note=ending_note)
 ```
 
 The basic `help` command will break commands up by cogs. Each cog will be a different page. Those pages can be navigated with
@@ -63,12 +66,13 @@ the arrow embeds. The message is unresponsive after 30s of no activity, it'll re
 
 # Changelog
 
-## [1.2.3]
-- Allowed for more customized ending notes.
+## [1.3.0]
+- Replaced `Navigation` with `DefaultMenu`
+- Allowed for custom Menus to be used, should inherit from `PrettyMenu` and must use `send_pages()` method
+- `active_time` is now part of the `DefaultMenu` args
 
 
 # Notes:
-
 - discord.py must already be installed to use this
 - `manage-messages` permission is recommended so reactions can be removed automatically
