+++ pretty_help/__init__.py
@@ -1,4 +1,4 @@
-from .navigation import *
+from .menu import *
 from .pretty_help import *
 
-__version__ = "1.2.3"
+__version__ = "1.3.0"
