+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "discord-pretty-help"
-version = "1.2.3"
+version = "1.3.0"
 description = "And nicer looking interactive help menu for discord.py"
 authors = ["StroupBSlayen <29642143+stroupbslayen@users.noreply.github.com>"]
 readme = "README.md"
